Keep stores as 2D arrays
We made guarantees about the processing order of stores. That
whenever an action is dispatched, stores will be processed in
the order with which they are added.

Unfortunately, before this commit we kept these stores in a POJO.
While many browsers treat objects as ordered maps _most_ of the time,
the cost to just use a 2D array is negligeable and allows us to drop
the `remap` module in favor of the more intuitive `Array.prototype.reduce`.

Additionally, it allows stores to be enumerated over using iterators like:

```
for (let [ key, store ] of app.stores) {}
```

Or using an ES6 Map like:

```
new Map(app.stores).forEach
```