Precompile transactions for 3-5x performance
Currently, the slowest part of the Microcosm dispatch process is the
invocation of the `register` method that Stores can implement to
subscribe to an action. This PR adds a registry to Microcosm instances
that memoizes this process. This cache is recalculated when a Microcosm
dispatches, and is invalidated whenever a new store is added to a
Microcosm instance.