Add concept of workflows
This commit adds a way to bundle together multiple actions into a common
"workflow". While a workflow is underway, it shares a common closure
with all subsequent workflow calls. This makes it easy to cancel prior
actions (such as outstanding networking requests for irrelevant
information).

Additionally, it simplifies Transaction in anticipation for a lower
level API for operating on transactions that will eventually be exposed
to the public.