Do not aggressively throw on missing actions
Before this commit, there was an error in the way stores report missing
handlers. They would check for `undefined` within the `register()`
function, however actions now have extra states that are referenced as
properties. We can no longer make this check when dispatching.

One alternative would be to use a Proxy. In the meantime, this commit
removes the warning and reduces another error to a warning.

Fixes #133