Eliminate internal Presenter wrapper shouldComponentUpdate
shouldComponentUpdate is implemented within a special context "sock"
inside of the presenter. This wrapper exists to ensure that React
context is properly transferred over when extending Presenters. IE10
or lower doesn't transfer static members for ES6 classes.

The reason I did this was to prevent the view model from recalculating
too many times when an app changes. If a parent is going to render
anyway, there's no sense in recalculating the model when the
presenter's subscription fires. Just do it all in the same render.

This commit removes that shouldComponentUpdate method and re-works the
efficiency test to test against the model recalculation, not
re-render.