Remove commit and shouldCommit
I've used commit on 3 projects, and I removed it on two of them (and
I'd like to remove it on all three).

Commit adds a lot of edge cases internally, and requires dependency on
other domains in order to be useful. For example, reducing down a
dataset to match a specific query.

We have indexes and computed props for that now, so the utility is no
longer there. Additionally, I want to remove as many use cases from
Domains that require external knowledge as possible.

I've also removed the ImmutableJS docs for the time being. When we
document indexing, we should dig into how to use them with ImmutableJS.