Pass command and status into register(). Command into intercept()
This commit updates the register method of Domains and Effects to
accept the given command and status for an action. Additionally, if a
function is returned from register, the dispatcher will use that
function as the handler.

This behavior also carries over to intercept(), however only the
command is given. This is because intercept applies _before_ an action
is created.