Action status methods are autobound
Action status methods, like `action.resolve()`, or `action.reject()`
are now auto-bound. This means that you can pass them directly into a
callback, and the scope of the method will remain as the action. This
is particularly useful when working with AJAX libraries. For example,
when working with `superagent`:

```javascript
import superagent from 'superagent'

function getPlanets () {
  return action => {
    let request = superagent.get('/planets')

    request.on('request', action.open)
    request.on('progress', action.update)

    request.then(action.resolve, action.reject)
  }
}
```

Related issues:
https://github.com/vigetlabs/microcosm/issues/144