Microcosms can be register to actions
Microcosms can now declare a custom register and getInitialState
method. In this way, they behave exactly like domains.

This does not allow Microcosms to be added as domains to other
Microcosms. That will require some work. This is intended to make it
easier to introduce Microcosm, and for concise examples.