Actions that push from presenters should use same repo
This commit fixes an issue where actions, that float to the top of the
presenter tree, are dispatched from the root repo, instead of the repo
for the presenter that dispatched them. This causes an unexpected
behavior where the root repo does not have state added by a domain
within a fork.

Additionally, before this commit, we had a check to compare the
histories of both repos within a Presenter hierarchy that contained
separate Microcosm history instances. We no longer need this check
because action dispatch occurs on the repo for the presenter that sent
the action.