Move Presenter batching to Model. Use History scheduler in Model.
I realized that the Preact Presenter wasn't updated with the batched
scheduling update from the React Presenter. This commit moves all of
that logic into the Model, adding a few optimizations along the way:

1. Models use the same update scheduler as history
2. Models do not queue an update if nothing has changed, so:
3. The Presenter mediator no longer needs to extend from PureComponent

For this to work, I had to make it so that the scheduler used by
History could be cancelled. For this, I borrowed some ideas from
https://github.com/yoshuawuyts/on-idle.