ActionForm and ActionButton actions should unsubscribe on mount
Before this commit, there was an issue where ActionForm and
ActionButton callbacks would still fire if the related components
unmounted. This resulted in an error where `setState` would call on an
unmounted component for most of the use cases for these callbacks.

This commit does a couple of things:

1. Action::execute: Moves coroutine behavior to the Action prototype
2. Actions know what repo they were created with. This is necessary
for execute, but could be used for other future work.
3. ActionQueue: A helper class for entities that must dispatch and
track multiple actions.

The result of this work is that ActionForm and ActionButton now
"empty" their queue of actions when they unmount, unsubscribing any
callbacks they set on those actions.

Long term, I think this could be particularly useful for Presenters to
know what actions they dispatched via `send`, but this is purely
speculative.