Remove compare tree
The CompareTree class was an addition added in the 12.x release of
Microcosm that made it easy to subscribe to fine-grained pathways
within a Microcosm state tree.

Usage in Microcosm was never publicly documented, and no internal
pieces rely on it. This commit moves this code into an add-on to avoid
"dead weight".

As we explore a querying interface for Microcosm, the CompareTree
might get more attention, however I want to separate it from Microcosm
core in the interim to free up some design decision making.