Domains are subjects
This commit changes the Domain class such that it extends from the
Subject prototype. Additionally, it configures a toJSON method that
serializes a domain according to history's HEAD state.

This means the `repo.answers` namespace can be removed. You can now
get updates to a domain simply by accessing the domain itself. This
required some Presenter tests to be updated.