[microcosm] Async, scheduled, Subjects
This commit brings the asychronous behavior of Observables over to
Subjects. It is also able to consolidate some behavior by extending
Observable.

The async behavior of Subjects is different in that they track a
present value. When subscribed to, Subjects immediately emit an
update. All future work is scheduled.

This is a really nice system because it allows Presenter view models
to immediately have a value, and for state to be synchronized by the
time Effects dispatch.