Pass next when completing with a payload. Improve efficiency
This commit allows Subjects to be completed with a payload. This
improves ergonomics and compatability with usage in Microcosm 12.x.

The drawback to this approach is that an extra action status is
dispatched. I have alleviated that by adding a `filter` method to the
Observable prototype that Domains use to avoid dispatch if an action
update has no impact on their state.

This actually ended up being a greater efficiency update than I
anticipated, allowing me to unskip a few tests.