Add back max history option
This commit adds back the max history option to Microcosm, a feature
of 12.x that controls how many actions Microcosm's history should hold
on to.

This is zero by default in order to prevent memory leaks. However some
apps (like email editors we have built in the past) may desire to hold
on to a set number of actions for undo/redo and reporting features.