Handle promise errors using second argument of then()
This commit changes the way Subjects and Microcosm action resolution
handle promises such that the error state is passed into the second
argument of `promise.then()`.

This prevents a case where callbacks following the success case could
raise an exception, but it would be trapped by the `.catch()` API call.